Version 1.3
Fully functional!
 Step 1. Connect to the wifi network called "Slider" with the password "123456789"
 Step 2. In a browser, go to 192.168.4.1
 Step 3. Press the "Calibrate" button (only necessary on the first startup)
 Step 4. Press the "Home" button
A webpage with a large HTML slider will appear. Moving this slider will
cause the motorized gantry to move in real-time. The arrow buttons on each end of the HTML slider allow for fine adjustment of its position. The start and end points can be set with the respective text fields, or by pressing "set to current". The speed can be set as a percentage (this appears by default) or as a time duration by first pressing "switch to time". Once the start point, end point, and time/speed are set, press the "run" button to make the slider perform the requested movement.